Adding check for undefined file id
Spent some time troubleshooting an issue with Shutterstock CMS where it would constantly fail and crash.
The issue was related to an item that was uploaded without a required field (thumbnail). Not
sure how this happened because the interface forces the user to upload a thumbnail. In any case
I think this check for an undefined file id should be added so that the application just doesnt
show a file rather than crash.